requirements.txt
django-ckeditor
django>=4.2
djangorestframework>=3.14
django-cors-headers>=4.3
gunicorn>=20.1
psycopg2-binary>=2.9
python-decouple>=3.8
whitenoise>=6.5
djangorestframework-simplejwt>=5.3
Pillow>=10.0.0
dj-database-url
django-cloudinary-storage
